major improvements

+++ run.sh
@@ -10,34 +10,70 @@ source /etc/fsl/5.0/fsl.sh #enable fsl (fsl6 still uses /etc/fsl/5.0.. for some 
 export PATH=$PATH:/usr/lib/mrtrix/bin #enable mrtrix
 export HOME=/ #so that tractseg uses /.tractseg not ~/.tractseg to look for prestaged models
 
-#opts=""
-#if [ $(jq -r .preprocess config.json) == "true" ]; then
-#	opts="$opts --preprocess"
-#fi
+opts=""
+if [ $(jq -r .preprocess config.json) == "true" ]; then
+	opts="$opts --preprocess"
+fi
 
-ln -sf $(jq -r .dwi config.json) dwi.nii.gz
-ln -sf $(jq -r .bvecs config.json) dwi.bvecs
-ln -sf $(jq -r .bvals config.json) dwi.bvals
+t1=`jq -r '.t1' config.json`
+nr_fibers=`jq -r '.nr_fibers' config.json`
 
-#t1=`jq -r '.t1' config.json`
-#if [ $t1 != "null" ]; then
-#	ln -sf $t1 T1w_acpc_dc_restore_brain.nii.gz
-#fi
+csd_type=`jq -r '.csd' config.json`
+if [ $csd_type == "csd_msmt_5tt" ]; then
+    if [ -f $t1 ]; then
+	    echo "Error: csd_msmt_5tt requires a T1w image."
+	    exit 1
+    else
+        ln -sf $t1 T1w_acpc_dc_restore_brain.nii.gz
+    fi
+    if [ $(jq -r .preprocess config.json) == "true" ]; then
+	    echo "Error: csd_msmt_5tt doesn't work with --preprocess."
+	    exit 1
+    fi
+fi
+
+bundles=`jq -r '.bundles' config.json`
+opts_bundles=""
+if [ "$bundles" != "" ]; then
+	opts_bundles="$opts_bundles --bundles $bundles"
+fi
 
 echo "(1/4) running tract_segmentation"
-TractSeg -i dwi.nii.gz --raw_diffusion_input \
-    --csd_type $(jq -r .csd config.json) \
-    --output_type tract_segmentation \
-    --keep_intermediate_files \
-    --nr_cpus 8 \
-    -o tractseg_output
+peaks=`jq -r '.peaks' config.json`
+if [ -f $peaks ]; then
+
+    echo "peaks.nii.gz found. Running TractSeg from peaks."
+    cp $(jq -r .peaks config.json) tractseg_output/peaks.nii.gz
+
+    TractSeg -i tractseg_output/peaks.nii.gz \
+        --output_type tract_segmentation \
+        --keep_intermediate_files \
+        --nr_cpus 8 \
+        -o tractseg_output 
+
+else
+
+    ln -sf $(jq -r .dwi config.json) dwi.nii.gz
+    ln -sf $(jq -r .bvecs config.json) dwi.bvecs
+    ln -sf $(jq -r .bvals config.json) dwi.bvals
+
+    TractSeg -i dwi.nii.gz --raw_diffusion_input \
+        --csd_type $(jq -r .csd config.json) \
+        --output_type tract_segmentation \
+        --keep_intermediate_files \
+        --nr_cpus 8 \
+        -o tractseg_output \
+        $opts
+           
+fi
 
 ##Get segmentations of the regions were the bundles start and end (helpful for filtering fibers that do not run from start until end).
 echo "(2/4) running endings_segmentation"
 TractSeg -i tractseg_output/peaks.nii.gz \
     --output_type endings_segmentation \
     --nr_cpus 8 \
-    -o tractseg_output
+    -o tractseg_output \
+    $opts
 
 #For each bundle create a Tract Orientation Map (Wasserthal et al., Tract orientation mapping for bundle-specific tractography). 
 #This gives you one peak per voxel telling you the main orientation of the respective bundle at this voxel. Can be used for 
@@ -48,20 +84,49 @@ echo "(3/4) running TOM/tracking"
 TractSeg -i tractseg_output/peaks.nii.gz \
     --output_type TOM \
     --nr_cpus 8 \
-    -o tractseg_output
+    -o tractseg_output \
+    $opts
 
 Tracking -i tractseg_output/peaks.nii.gz \
     --tracking_format tck \
     --nr_cpus 8 \
-    -o tractseg_output
+    --nr_fibers $nr_fibers \
+    -o tractseg_output \
+    $opts_bundles
 
 echo "(4/4) running Tractometry"
-Tractometry -i tractseg_output/TOM_trackings/ \
-    -o tractseg_output/Tractometry_peaks.csv \
-    -e tractseg_output/endings_segmentations/ \
-    -s tractseg_output/peaks.nii.gz \
-    --tracking_format tck \
-    --TOM tractseg_output/TOM \
-    --peak_length
+#By default, tractometry is run over the peak_length, since it doesn't require the tensor as input.
+#However, if the tensor is given as input, the user may decide to run tractometry on either FA, MD, RD, or AD.
+
+tractometry_input=`jq -r '.tractometry_input' config.json`
+if [ $tractometry_input == 'peak_length' ]; then
+
+    Tractometry -i tractseg_output/TOM_trackings/ \
+        -o tractseg_output/Tractometry_peaks.csv \
+        -e tractseg_output/endings_segmentations/ \
+        -s tractseg_output/peaks.nii.gz \
+        --tracking_format tck \
+        --TOM tractseg_output/TOM \
+        --peak_length
+
+else 
+
+    fa=`jq -r '.fa' config.json`
+    ad=`jq -r '.ad' config.json`
+    md=`jq -r '.md' config.json`
+    rd=`jq -r '.rd' config.json`
+    image=${!tractometry_input}
+    
+    if [ -f $image ]; then 
+        Tractometry -i tractseg_output/TOM_trackings/ \
+            -o tractseg_output/Tractometry_peaks.csv \
+            -e tractseg_output/endings_segmentations/ \
+            -s $image \
+            --tracking_format tck 
+    else
+        echo "Error: $image does not exist. Check your tensor input."
+        exit 1
+    fi
+fi    
 
 echo "all done with tractseg"
